Rename azle_vm_value_derive crate
