+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
 members = [
     "src/compiler/typescript_to_rust/azle_generate",
-    "src/compiler/typescript_to_rust/azle_js_value_derive"
+    "src/compiler/typescript_to_rust/azle_vm_value_derive"
 ]
