+++ src/compiler/typescript_to_rust/azle_generate/src/cdk_act/nodes/data_type_nodes/act_record.rs
@@ -96,7 +96,7 @@ impl ToTokenStream for RecordTypeAlias {
             .map(|member| member.to_token_stream())
             .collect();
         quote!(
-            #[derive(serde::Deserialize, Debug, candid::CandidType, Clone, CdkActTryIntoVmValue, TryFromVmValue)]
+            #[derive(serde::Deserialize, Debug, candid::CandidType, Clone, CdkActTryIntoVmValue, CdkActTryFromVmValue)]
             struct #type_ident {
                 #(#member_token_streams),*
             }
