+++ src/compiler/typescript_to_rust/azle_generate/src/cdk_act/nodes/data_type_nodes/act_tuple.rs
@@ -84,7 +84,7 @@ impl ToTokenStream for TupleTypeAlias {
             .map(|elem| elem.to_token_stream())
             .collect();
         quote!(
-            #[derive(serde::Deserialize, Debug, candid::CandidType, Clone, CdkActTryIntoVmValue, TryFromVmValue)]
+            #[derive(serde::Deserialize, Debug, candid::CandidType, Clone, CdkActTryIntoVmValue, CdkActTryFromVmValue)]
             struct #type_ident (
                 #(#elem_idents),*
             );
