+++ src/compiler/typescript_to_rust/azle_generate/src/cdk_act/nodes/data_type_nodes/act_variants.rs
@@ -83,7 +83,7 @@ impl ToTokenStream for VariantTypeAlias {
             .map(|member| member.to_token_stream())
             .collect();
         quote!(
-            #[derive(serde::Deserialize, Debug, candid::CandidType, Clone, CdkActTryIntoVmValue, TryFromVmValue)]
+            #[derive(serde::Deserialize, Debug, candid::CandidType, Clone, CdkActTryIntoVmValue, CdkActTryFromVmValue)]
             enum #type_ident {
                 #(#member_token_streams),*
             }
